update preview routes to folder/file structure
